Fix: add missing animateFloat import + harden sin/coerceIn types
Same root cause as the vinyl build failure: forgot the
androidx.compose.animation.core.animateFloat import (the
InfiniteTransition extension). Once that's unresolved, wavePhase
has no type, which cascades into 6 follow-on errors on the +, sin,
and coerceIn calls.

Also hardened the sin/coerceIn calls to use Double math explicitly
(toDouble(), Double literals) to avoid future overload ambiguity
between sin(Float) and sin(Double). Convert to Float at the very
end.

Errors fixed:
  QuickPicksScreen.kt:176:41 Unresolved reference 'animateFloat'
  QuickPicksScreen.kt:179:25 Cannot infer type for this parameter
  QuickPicksScreen.kt:180:25 Cannot infer type for this parameter
  QuickPicksScreen.kt:367:55 Overload resolution ambiguity
  QuickPicksScreen.kt:368:71 Float vs Double mismatch
  QuickPicksScreen.kt:368:77 Float vs Double mismatch
  QuickPicksScreen.kt:374:43 Float was expected

+++ app/src/main/java/com/rajatxo/coral/ui/screens/QuickPicksScreen.kt
@@ -5,6 +5,7 @@ import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.RepeatMode
 import androidx.compose.animation.core.animateDpAsState
+import androidx.compose.animation.core.animateFloat
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.infiniteRepeatable
 import androidx.compose.animation.core.rememberInfiniteTransition
@@ -364,8 +365,12 @@ fun QuickPicksScreen(
                             val layerOpacity = if (isActive) {
                                 1f
                             } else {
-                                val phase = wavePhase + i * 0.4f
-                                (0.55f + 0.18f * sin(phase)).coerceIn(0.4f, 0.78f)
+                                // Use Double math throughout to avoid sin(Float) vs sin(Double)
+                                // overload ambiguity, then convert to Float at the end.
+                                val phase = wavePhase.toDouble() + i * 0.4
+                                (0.55 + 0.18 * sin(phase))
+                                    .coerceIn(0.4, 0.78)
+                                    .toFloat()
                             }
 
                             AbyssalLayerRow(
